Only guess column types from the first 2k rows.  return an error string for errors in parser iteration.
Guesses were wrong on full-dataset scans anyway - floats with too much precision, for example.  So, save a ton of time on big datasets by not processing the whole file.

Also - sometimes casting a value to the column type value can fail, while the parser is being iterated.  Consumers should check for a string instead of a list.